requirements.txt
anyio==3.6.1
attrs==22.1.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
peewee==3.15.1
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.2
